// src/app/usuarios/payload-obj.interface.ts
export interface PayloadObj {
    info_adicional: string;
    user_name:      string;
    scope:          string[];
    apellido:       string;
    exp:            number;
    nombre:         string;
    authorities:    string[];
    jti:            string;
    email:          string;
    client_id:      string;
}